requirements.txt
acme==2.6.0
adal==1.2.7
aiobotocore==2.5.4
aiohttp==3.7.4.post0
aioitertools==0.7.1
altgraph==0.17
arnparse==0.0.2
asgiref==3.4.1
asn1crypto==1.4.0
async-timeout==3.0.1
attr==0.3.1
attrs==21.2.0
autoflake==1.5.3
awswrangler==3.0.0
azure-common==1.1.28
azure-core==1.29.3
azure-cosmos==4.3.0
azure-graphrbac==0.61.1
azure-identity==1.12.0
azure-keyvault-keys==4.8.0
azure-mgmt-appcontainers==2.0.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.0.0
azure-mgmt-keyvault==10.2.1
azure-mgmt-loganalytics==12.0.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-msi==7.0.0
azure-mgmt-network==23.1.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==22.0.0
azure-mgmt-sql==3.0.1
azure-mgmt-storage==21.0.0
azure-mgmt-subscription==3.1.1
azure-mgmt-web==7.1.0
azure-storage-blob==12.17.0
beautifulsoup4==4.9.3
black==22.1.0
boto3==1.28.1
botocore==1.31.17
cachelib==0.9.0
cachetools==5.3.1
certbot==2.5.0
certbot-nginx==2.5.0
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
click==8.0.3
coloredlogs==15.0
component-extraction-api==0.0.1
ConfigArgParse==1.7
configobj==5.0.8
configparser==5.0.2
cryptography==38.0.1
cx-Oracle==8.3.0
Deprecated==1.2.13
distro==1.8.0
dj-git==0.3
Django==3.2.5
docopt==0.6.2
ecdsa==0.14.1
elementpath==3.0.2
et-xmlfile==1.1.0
fake-awsglue==0.0.0.post20190320
findspark==2.0.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Session==0.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.36
google-api-core==2.11.1
google-api-python-client==2.86.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
humanfriendly==9.1
idna==2.10
importchecker==2.0
importlib-metadata==4.11.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.0.1
jmespath==0.10.0
josepy==1.13.0
json-flatten==0.1
kafka-python==2.0.2
lambda-git==0.1.1
lark-parser==0.10.1
lxml==4.9.2
macholib==1.14
Markdown==3.4.1
MarkupSafe==2.0.1
modules-for-mozia==1.1.0
mozia-modules==1.8.3
msal==1.23.0
msal-extensions==1.0.0
msgraph-core==0.2.2
msrest==0.7.1
msrestazure==0.6.4
multidict==5.1.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
names==0.3.0
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.0.7
pandas==1.5.3
parsedatetime==2.6
pathspec==0.9.0
pg8000==1.19.5
pip-autoremove==0.10.0
pipdeptree==2.2.0
pipreqs==0.4.13
platformdirs==2.5.0
portalocker==2.7.0
progress==1.5
proto-plus==1.22.3
protobuf==4.25.0
psycopg2-binary==2.9.6
public-ip==0.12
py4j==0.10.9
pyarmor==6.7.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
pyathena==2.2.0
pycparser==2.20
pyfiglet==1.0.1
pyflakes==2.5.0
pygal==2.4.0
pygit2==1.12.0
pyhcl==0.4.4
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2021.5
PyJWT==2.8.0
pymssql==2.2.7
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.1.1
pyRFC3339==1.1
pyspark==3.1.1
python-dateutil==2.8.1
python-dotenv==0.21.1
python-hcl2==3.0.5
python-jose==3.2.0
pytz==2021.1
PyYAML==6.0
redis==3.5.3
redis-py-cluster==2.1.2
redshift-connector==2.0.882
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3fs==2023.6.0
s3transfer==0.6.0
scramp==1.4.0
six==1.16.0
smmap==5.0.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
sqlparse==0.4.1
stripe==4.1.0
tabulate==0.8.9
tenacity==7.0.0
tinyaes==1.0.4
tomark==0.1.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.4
uWSGI==2.0.21
Werkzeug==2.1.2
wrapt==1.12.1
yarg==0.1.9
yarl==1.9.2
zipp==3.8.0
